Revert "Robot driving"
